requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# 图像处理
Pillow>=8.3.0
opencv-python>=4.5.0
albumentations>=1.3.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 进度条
tqdm>=4.62.0

# Jupyter支持
jupyter>=1.0.0
ipykernel>=6.0.0

# 其他工具
requests>=2.26.0
pathlib2>=2.3.0

# 自然语言处理
transformers>=4.20.0
tokenizers>=0.12.0
datasets>=2.0.0

# 模型训练工具
tqdm>=4.62.0
wandb>=0.12.0
tensorboardX>=2.4.0

# 数据增强
albumentations>=1.2.0
imgaug>=0.4.0

# 工具库
requests>=2.25.0
pyyaml>=5.4.0
click>=8.0.0
rich>=12.0.0

# 开发工具
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
isort>=5.9.0

# 可选：CUDA支持 (根据系统选择)
# torch-audio  # 如果需要音频处理
# torchtext    # 如果需要文本处理

# 可选：高级功能
# onnx>=1.12.0           # 模型转换
# onnxruntime>=1.12.0    # ONNX运行时
# tensorrt               # NVIDIA TensorRT (需要NVIDIA GPU)